Correct failing tests in the m/3d/imagedraw class.
On armhf, arm64, and ppc64el a static cast from float to unsigned
actually does a proper range checking and maps negative values to
zero. Hence one first has to cast to a signed integer value and
then to the unsigned version in order to save one comparison.

+++ mia/3d/imagedraw.cc
@@ -43,9 +43,9 @@ C3DDrawBox::C3DDrawBox(const C3DBounds& size, const C3DFVector& origin, const C3
 void C3DDrawBox::draw_point(const C3DFVector& p)
 {
        
-        C3DBounds ip(static_cast<unsigned>(roundf(p.x)), 
-                     static_cast<unsigned>(roundf(p.y)), 
-                     static_cast<unsigned>(roundf(p.z)));
+        C3DBounds ip(static_cast<short>(roundf(p.x)), 
+                     static_cast<short>(roundf(p.y)), 
+                     static_cast<short>(roundf(p.z)));
        
 	cvdebug() << "about to draw " << ip << " from " << p << "\n"; 
         if (ip < m_size)
